refactor: strip -ts folder suffix
