+++ src/core/GeometryRepresentation.tsx
@@ -1,12 +1,10 @@
-import vtkActor2D, {
-  IActor2DInitialValues,
-} from '@kitware/vtk.js/Rendering/Core/Actor2D';
-import { ICoordinateInitialValues } from '@kitware/vtk.js/Rendering/Core/Coordinate';
-import { Coordinate } from '@kitware/vtk.js/Rendering/Core/Coordinate/Constants';
-import vtkMapper2D, {
-  IMapper2DInitialValues,
-} from '@kitware/vtk.js/Rendering/Core/Mapper2D';
-import { IProperty2DInitialValues } from '@kitware/vtk.js/Rendering/Core/Property2D';
+import vtkActor, {
+  IActorInitialValues,
+} from '@kitware/vtk.js/Rendering/Core/Actor';
+import vtkMapper, {
+  IMapperInitialValues,
+} from '@kitware/vtk.js/Rendering/Core/Mapper';
+import { IPropertyInitialValues } from '@kitware/vtk.js/Rendering/Core/Property';
 import { Vector2 } from '@kitware/vtk.js/types';
 import {
   forwardRef,
@@ -17,16 +15,15 @@ import {
   useState,
 } from 'react';
 import { IDownstream, IRepresentation } from '../types';
-import { compareShallowObject } from '../utils-ts/comparators';
-import useBooleanAccumulator from '../utils-ts/useBooleanAccumulator';
-import useComparableEffect from '../utils-ts/useComparableEffect';
+import { compareShallowObject } from '../utils/comparators';
+import useBooleanAccumulator from '../utils/useBooleanAccumulator';
+import useComparableEffect from '../utils/useComparableEffect';
 import {
   DownstreamContext,
   RepresentationContext,
   useRendererContext,
 } from './contexts';
 import useColorTransferFunction from './modules/useColorTransferFunction';
-import useCoordinate from './modules/useCoordinate';
 import useMapper from './modules/useMapper';
 import useProp from './modules/useProp';
 
@@ -39,17 +36,17 @@ interface Props extends PropsWithChildren {
   /**
    * Properties to set to the actor
    */
-  actor?: IActor2DInitialValues;
+  actor?: IActorInitialValues;
 
   /**
    * Properties to set to the actor
    */
-  mapper?: IMapper2DInitialValues;
+  mapper?: IMapperInitialValues;
 
   /**
    * Properties to set to the actor.property
    */
-  property?: IProperty2DInitialValues;
+  property?: IPropertyInitialValues;
 
   /**
    * Preset name for the lookup table color map
@@ -62,20 +59,43 @@ interface Props extends PropsWithChildren {
   colorDataRange?: [number, number];
 
   /**
-   * The coordinate system in which the input dataset resides.
+   * Show/Hide Cube Axes for the given representation
    */
-  transformCoordinate?: ICoordinateInitialValues;
+  showCubeAxes?: boolean;
+
+  /**
+   * Configure cube Axes style by overriding the set of properties defined
+   * https://github.com/Kitware/vtk-js/blob/HEAD/Sources/Rendering/Core/CubeAxesActor/index.js#L703-L719
+   *
+   * TODO fix type
+   */
+  cubeAxesStyle?: Record<string, unknown>;
+
+  /**
+   * Show hide scalar bar for that representation
+   */
+  showScalarBar?: boolean;
+
+  /**
+   * Use given string as title for scalar bar. By default it is empty (no title).
+   */
+  scalarBarTitle?: boolean;
+
+  /**
+   * Configure scalar bar style by overriding the set of properties defined
+   * https://github.com/Kitware/vtk-js/blob/master/Sources/Rendering/Core/ScalarBarActor/index.js#L776-L796
+   *
+   * TODO fix type
+   */
+  scalarBarStyle?: Record<string, unknown>;
 }
 
 const DefaultProps = {
   colorMapPreset: 'erdc_rainbow_bright',
   colorDataRange: [0, 1] as Vector2,
-  transformCoordinate: {
-    coordinateSystem: Coordinate.DISPLAY,
-  },
 };
 
-export default forwardRef(function Geometry2DRepresentation(
+export default forwardRef(function GeometryRepresentation(
   props: Props,
   fwdRef
 ) {
@@ -90,24 +110,21 @@ export default forwardRef(function Geometry2DRepresentation(
     trackModified
   );
 
-  // --- coordinate --- //
-
-  const { transformCoordinate = DefaultProps.transformCoordinate } = props;
-  const getCoordinate = useCoordinate(transformCoordinate, trackModified);
-
   // --- mapper --- //
 
-  const getMapper = useMapper<vtkMapper2D, IMapper2DInitialValues>(
+  const getMapper = useMapper<vtkMapper, IMapperInitialValues>(
     () =>
-      vtkMapper2D.newInstance({
+      vtkMapper.newInstance({
         lookupTable: getLookupTable(),
-        useLookupTableScalarRange: false,
-        scalarVisibility: false,
-        transformCoordinate: getCoordinate(),
-      } as IMapper2DInitialValues),
+        useLookupTableScalarRange: true,
+      } as IMapperInitialValues),
     props.mapper,
     trackModified
   );
+
+  useEffect(() => {
+    getMapper().setLookupTable(getLookupTable());
+  }, [getMapper, getLookupTable]);
 
   // --- actor --- //
 
@@ -115,19 +132,15 @@ export default forwardRef(function Geometry2DRepresentation(
     ...props.actor,
     visibility: dataAvailable && (props.actor?.visibility ?? true),
   };
-  const getActor = useProp<vtkActor2D, IActor2DInitialValues>({
-    constructor: () => vtkActor2D.newInstance({ visibility: false }),
+  const getActor = useProp<vtkActor, IActorInitialValues>({
+    constructor: () => vtkActor.newInstance({ visibility: false }),
     id: props.id,
     props: actorProps,
     trackModified,
   });
 
   useEffect(() => {
-    // TODO type hack; upgrade vtk.js
-    const actor = getActor() as vtkActor2D & {
-      setMapper(mapper: vtkMapper2D): boolean;
-    };
-    actor.setMapper(getMapper());
+    getActor().setMapper(getMapper());
   }, [getActor, getMapper]);
 
   // set actor property props
