+++ src/core/View.tsx
