+++ src/core/internal/view-shared.ts
