+++ src/core/modules/useApplyCenterOfRotation.ts
