+++ src/utils/DeletionRegistry.ts
