+++ src/utils/useBooleanAccumulator.ts
